requirements.txt
# Core Dependencies
python>=3.8

# NLP and Text Processing
spacy>=3.4.0
# After installation, run: python -m spacy download en_core_web_sm

# PDF Processing (FREE - Open Source)
pdfplumber>=0.9.0  # Better text extraction from PDFs
PyPDF2>=3.0.0      # Alternative PDF reader (backup)

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.3.0  # For TF-IDF and cosine similarity

# Configuration
pyyaml>=6.0

# String Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Optional: For semantic similarity (lightweight)
sentence-transformers>=2.2.0

# Web UI
streamlit>=1.25.0

# API Integration
requests>=2.31.0

# Optional: AI Enhancements
# openai>=1.0.0  # Uncomment if using OpenAI
# google-generativeai>=0.3.0  # Uncomment if using Gemini
python-dotenv>=1.0.0  # For environment variables

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0

